feat: 修改变量命名
