+++ src/components/video/video.less
@@ -21,7 +21,7 @@
     bottom: 0;
     width: 100%;
     height: 60px;
-    background: linear-gradient(#ccc,rgba(0,0,0,0.5));
+    background: linear-gradient(rgba(204, 204, 204,0.5),rgba(0,0,0,0.5));
     transition: all .75s;
     opacity: 1;
     z-index: 133;
@@ -37,7 +37,7 @@
       align-items: center;
       .antd-waffle-video-play-line{
         width:calc(100% - 84px);
-        background-color:rgba(204, 204, 204, .5);
+        background-color:rgba(205, 205, 205, .5);
         border-radius: 2px;
         height: 2px;
         position: relative;
